requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
chromadb
rank_bm25
python-dotenv
tiktoken
pysqlite3-binary
pandas
nltk